+++ src/index.js
@@ -1,2 +1,2 @@
-export { default as default } from "./Accordion.svelte";
+export { default as Accordion } from "./Accordion.svelte";
 export { default as AccordionItem } from "./AccordionItem.svelte";