+++ test.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import { Accordion, AccordionItem } from "./types";
+</script>
+
+<Accordion multiselect>
+  <AccordionItem expanded title="Title 1">Content 1</AccordionItem>
+  <AccordionItem title="Title 2">Content 2</AccordionItem>
+  <AccordionItem title="Title 3">Content 2</AccordionItem>
+</Accordion>