+++ types/index.d.ts
@@ -1,2 +1,2 @@
-export { default as default } from "./Accordion";
+export { default as Accordion } from "./Accordion";
 export { default as AccordionItem } from "./AccordionItem";